requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-rest-auth==2.2.4
Django==4.0.4
django-jalali==6.0.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
jdatetime==4.1.0
Pillow==9.1.0
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
